requirements.txt
django
openpyxl
pandas
djangorestframework
markdown
django-filter
xlsxwriter
xlrd
django-cors-headers